requirements.txt
flask
flask-socketio
pymongo[srv]
python-dotenv
eventlet
flask-cors
certifi
